split behat ui tests into new lpa context file
